update(bashrc): aliases, move function into script
Adds some new aliases, updates some of my git aliases to have better
flags.

update(btc): move into a script

Move my btc function into a script.
It's still a fucntion in there, but I don't want it in my bashrc anymore
To use it, it has to be sourced first with `source path/to/btc`