add(laptop): bash_logout file
Prints '[ Goodbye! ]' with figlet when I logout. Also sleeps for 0.5
seconds. I had an idea to make it read a variable to say whether or not
to sleep, if I was trying to do a timed reboot or something, but if I
know I want to do that I'll just comment it out or something.