feat(network bar): cleanup/extra functionality
Restart dhcpcd when clicking ethernet modules (determined by first
letter: w or e)
Define new usage rules: set [name] to a descriptive tag or the interface
to use, and define command in the block. This is to allow multiple
network blocks with different names, and will use those names
constructively in the script.
Also make a good comment describing variables, usage, and dependencies
Clean up some flow in the script as well to make things nice
All around, good update to this module